[FIX] Accept coderabbit's suggest
주요 변경사항은 다음과 같습니다.
1. 테스트 코드의 구조도 리팩토링한 구조와 동일하게 수정했습니다.
2. utiliry -> utility 로 오타 수정했습니다.
3. 구현체를 참조하던 코드를 인터페이스를 참조하도록 수정했습니다.
